Unify interface of stepper variants
Currently, the shared component module serves two stepper variants. In the first step, both interfaces get aligned, remove unnecessary dependencies like translation. In a further refactoring step, both variants have to merge into one.
