+++ docs/src/pages/pidp/approach/byExample/use.tsx
@@ -43,7 +43,13 @@ const generateContent = (t: Translate): Array<Content> => {
 
 const use = () => {
   const { t } = useTranslation();
-  return <Stepper.StepperContent elements={generateContent(t)} />;
+  return (
+    <Stepper.StepperContent
+      elements={generateContent(t)}
+      labelBack={t('common:back')}
+      labelNext={t('common:next')}
+    />
+  );
 };
 
 export default use;
