+++ packages/components/package.json
@@ -10,7 +10,6 @@
     "@material-ui/icons": "*",
     "clsx": "*",
     "next": "*",
-    "next-translate": "*",
     "react": "*",
     "react-dom": "*",
     "use-resize-observer": "*"
@@ -24,7 +23,6 @@
     "@types/react-dom": "^17.0.0",
     "clsx": "^1.1.1",
     "next": "^10.0.4-canary.3",
-    "next-translate": "1.0.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "typescript": "^4.1.2",
