+++ packages/components/src/stepper/StepperContentAdv.tsx
@@ -1,9 +1,8 @@
 import { CardContent, createStyles, Grid, makeStyles, Typography } from '@material-ui/core';
-import useTranslation from 'next-translate/useTranslation';
-import React, { useState } from 'react';
+import React, { FC, useState } from 'react';
 
 import { StepperAdv } from './StepperAdv';
-import { Content, Stack } from './types';
+import { Content, MergedStepperContentProps } from './types';
 
 const useStyles = makeStyles(() =>
   createStyles({
@@ -52,9 +51,11 @@ export const renderTitleAndDescription = (items: Array<Content> = []) => {
   });
 };
 
-export const StepperContentAdv = ({ elements = [] }: Stack) => {
-  const { t } = useTranslation();
-
+export const StepperContentAdv: FC<MergedStepperContentProps> = ({
+  elements = [],
+  labelBack,
+  labelNext
+}) => {
   const [step, setStep] = useState(0);
 
   const max = getStepsLength(elements);
@@ -69,8 +70,8 @@ export const StepperContentAdv = ({ elements = [] }: Stack) => {
           setStepCb={(currentStep: number) => {
             setStep(currentStep);
           }}
-          labelNext={t('common:next')}
-          labelBack={t('common:back')}
+          labelBack={labelBack}
+          labelNext={labelNext}
         />
       </Grid>
       {renderTitleAndDescription(currentSteps)}
